Style the header and add new content
In this commit i styled the header the way I wanted. There were some
tricky parts, like the body having a default margin, but I was able to
figure it out.

I also added some content for the next part of the website, which I
shouldn't have in this commit. I ended up forgetting to commit the style
commit, which should not happen