Add call to action content
This commit was to add call to action part of the page. I added some
style as well from the classes that already have some ,such as "text"
and "buttons". I also added a style for the border of the button. Will
see if this will be taken into consideration or not when I add the
remaining style.