Add correct images and their sizes
In this commit I included the images I wanted to use on the website.
I needed to tweak the sizes to make them all fit together nicely.

Only thing is, my website is not responsive at all, the web page is
hideous. I know how to fix it, but need to work with the "flex"
css to iron it out. Will do in the future but for now, the webpage is
finished